server/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.2.0
Click==7.0
colorama==0.4.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.7
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.4.0
duplicity==0.8.12.0
email-validator==2.0.0.post2
entrypoints==0.3
exceptiongroup==1.1.2
fastapi==0.100.0
fasteners==0.14.1
filelock==3.12.2
frozenlist==1.4.0
future==0.18.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httplib2==0.14.0
httptools==0.6.0
httpx==0.24.1
idna==2.8
itsdangerous==2.1.2
Jinja2==3.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.3
monotonic==1.5
multidict==6.0.4
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
openai==0.27.8
orjson==3.9.2
paramiko==2.6.0
pbr==5.11.1
pexpect==4.6.0
Pillow==7.0.0
platformdirs==3.9.1
protobuf==3.6.1
psycopg2-binary==2.9.6
pycairo==1.16.2
pycups==1.9.73
pydantic==2.0.3
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.3.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
systemd-python==234
testresources==2.0.1
tqdm==4.65.0
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==5.8.0
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
uvicorn==0.23.1
uvloop==0.17.0
virtualenv==20.24.1
wadllib==1.3.3
watchfiles==0.19.0
websockets==11.0.3
xkit==0.0.0
yarl==1.9.2
